rename userSlice to authSlice for clarity
