--- client/src/store/userSlice/index.ts
@@ -1,6 +1,0 @@
-import userReducer, { login, selectUserName, selectUser, selectUserState } from './user.slice';
-import { HaztrakUser } from './user.slice';
-
-export default userReducer;
-export { login, selectUserName, selectUser, selectUserState };
-export type { HaztrakUser };
